pyproject.toml
[tool.poetry]
name = "apiruns"
version = "0.0.4"
description = "Apiruns CLI is a tool to make self-configurable rest API."
authors = ["Jose Salas <jose.salas@apiruns.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/apiruns/apiruns-cli"
repository = "https://github.com/apiruns/apiruns-cli"
documentation = "https://github.com/apiruns/apiruns-cli"
keywords = [
    "Poetry",
    "CLI",
    "Apiruns",
    "Packages",
    "Packaging",
    "Docker",
    "No-code",
    "Low-code",
]

[tool.poetry.dependencies]
python = "^3.7"
typer = {extras = ["all"], version = "^0.6.1"}
Cerberus = "^1.3.4"
requests = "^2.28.1"
httpx = "^0.23.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
apiruns = "apiruns.main:app"
